requirements_dev.txt
black==19.3b0
Django==2.2
django-extensions==2.1.6
djangorestframework==3.9.2
factory-boy==2.11.1
pylama==7.7.1